Refactor StringMatcher and children to dedupe code
