+++ sdk/scheduler/src/main/java/com/mesosphere/sdk/offer/evaluate/placement/AttributeRule.java
@@ -4,8 +4,6 @@ import java.util.Collection;
 import java.util.stream.Collectors;
 
 import com.mesosphere.sdk.specification.PodInstance;
-import org.apache.commons.lang3.builder.EqualsBuilder;
-import org.apache.commons.lang3.builder.HashCodeBuilder;
 import org.apache.mesos.Protos.Attribute;
 import org.apache.mesos.Protos.Offer;
 import org.apache.mesos.Protos.TaskInfo;
@@ -21,48 +19,26 @@ import com.fasterxml.jackson.annotation.JsonProperty;
  * @see AttributeStringUtils#toString(Attribute)
  */
 public class AttributeRule extends StringMatcherRule {
-    private final StringMatcher matcher;
-
     @JsonCreator
     public AttributeRule(@JsonProperty("matcher") StringMatcher matcher) {
-        this.matcher = matcher;
+        super("AttributeRule", matcher);
     }
 
     @Override
     public EvaluationOutcome filter(Offer offer, PodInstance podInstance, Collection<TaskInfo> tasks) {
-        if (isAcceptable(matcher, offer, podInstance, tasks)) {
+        if (isAcceptable(offer, podInstance, tasks)) {
             return EvaluationOutcome.pass(
                     this,
-                    "Match found for attribute pattern: '%s'", matcher.toString())
+                    "Match found for attribute pattern: '%s'", getMatcher().toString())
                     .build();
         } else {
             return EvaluationOutcome.fail(
                     this,
                     "None of %d attributes matched pattern: '%s'",
                     offer.getAttributesCount(),
-                    matcher.toString())
+                    getMatcher().toString())
                     .build();
         }
-    }
-
-    @JsonProperty("matcher")
-    private StringMatcher getMatcher() {
-        return matcher;
-    }
-
-    @Override
-    public String toString() {
-        return String.format("AttributeRule{matcher=%s}", matcher);
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        return EqualsBuilder.reflectionEquals(this, o);
-    }
-
-    @Override
-    public int hashCode() {
-        return HashCodeBuilder.reflectionHashCode(this);
     }
 
     @Override
