+++ sdk/scheduler/src/main/java/com/mesosphere/sdk/offer/evaluate/placement/StringMatcherRule.java
@@ -1,6 +1,9 @@
 package com.mesosphere.sdk.offer.evaluate.placement;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
 import com.mesosphere.sdk.specification.PodInstance;
+import org.apache.commons.lang3.builder.EqualsBuilder;
+import org.apache.commons.lang3.builder.HashCodeBuilder;
 import org.apache.mesos.Protos;
 
 import java.util.Collection;
@@ -10,15 +13,39 @@ import java.util.Collection;
  * of some key (e.g. attribute, hostname, region, zone ...).
  */
 public abstract class StringMatcherRule implements PlacementRule {
+    private final StringMatcher matcher;
+    private final String name;
+
+    protected StringMatcherRule(String name, StringMatcher matcher) {
+        this.name = name;
+        this.matcher = matcher;
+    }
+
+    @JsonProperty("matcher")
+    protected StringMatcher getMatcher() {
+        return matcher;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        return EqualsBuilder.reflectionEquals(this, o);
+    }
+
+    @Override
+    public int hashCode() {
+        return HashCodeBuilder.reflectionHashCode(this);
+    }
+
+    @Override
+    public String toString() {
+        return String.format("%s{matcher=%s}", name, matcher);
+    }
+
     public abstract Collection<String> getKeys(Protos.Offer offer);
 
-    public boolean isAcceptable(
-            StringMatcher matcher,
-            Protos.Offer offer,
-            PodInstance podInstance,
-            Collection<Protos.TaskInfo> tasks) {
+    protected boolean isAcceptable(Protos.Offer offer, PodInstance podInstance, Collection<Protos.TaskInfo> tasks) {
         for (String key : getKeys(offer)) {
-            if (matcher.matches(key)) {
+            if (getMatcher().matches(key)) {
                 return true;
             }
         }
